+++ Dockerfile-dev
@@ -18,10 +18,10 @@ RUN python3 -m venv --copies /app/venv; \
     poetry install --no-root --no-interaction --with dev
 ENV PATH /app/venv/bin:$PATH
 
-COPY hello_world hello_world
+COPY MPSSRC src
 COPY tests tests
 
 ENV ENV dev
 ENV APP_ENV $ENV
 
-CMD ["poetry", "run", "pytest", "--cov-report", "term-missing", "--cov=hello_world", "--typeguard-packages=hello_world"]
+CMD ["poetry", "run", "pytest", "--cov-report", "term-missing", "--cov=src", "--typeguard-packages=src"]
