+++ MPSSRC/__init__.py
@@ -1,0 +1,1 @@
+"""The main module of the MPSSRC package."""
