+++ docker-compose-dev.yml
@@ -1,8 +1,8 @@
 version: '3.8'
 
 services:
-  hello_world_app:
+  MPSSRC_app:
     build:
         context: .
         dockerfile: Dockerfile-dev
-    container_name: hello_world_app_dev
+    container_name: MPSSRC_app_dev
