--- hello_world/__init__.py
@@ -1,1 +1,0 @@
-"""The main module of the hello_world package."""
