+++ tests/hello_world_test.py
@@ -1,6 +1,6 @@
 from pytest import LogCaptureFixture, raises
 
-from hello_world.hello_world import get_hello_world
+from MPSSRC.hello_world import get_hello_world
 
 
 def test_hello_world(
